add authUser to invoice controller with error messages

+++ src/main/java/com/lucaapps/server/domain/invoice/services/InvoiceServiceImpl.java
@@ -1,5 +1,6 @@
 package com.lucaapps.server.domain.invoice.services;
 
+import com.lucaapps.server.domain.user.entities.AppUser;
 import com.lucaapps.server.exception.AppException;
 import com.lucaapps.server.exception.Error;
 import com.lucaapps.server.domain.invoice.dtos.InvoiceDto;
@@ -10,6 +11,7 @@ import com.lucaapps.server.domain.invoice.entities.Invoice;
 import com.lucaapps.server.domain.invoice.entities.Item;
 import com.lucaapps.server.domain.invoice.repository.InvoiceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -29,8 +31,8 @@ public class InvoiceServiceImpl implements InvoiceService {
 
 
     @Override
-    public List<InvoiceDto> getAllInvoices() {
-        List<Invoice> invoices = this.invoiceRepository.findAll();
+    public List<InvoiceDto> getAllInvoices(AppUser authUser) {
+        List<Invoice> invoices = this.invoiceRepository.findInvoicesByUserId(authUser.getId());
         List<InvoiceDto> invoiceDtos = new ArrayList<>();
         for (Invoice i : invoices) {
             invoiceDtos.add(
@@ -45,19 +47,13 @@ public class InvoiceServiceImpl implements InvoiceService {
         return invoiceDtos;
     }
 
-    @Override
-    public Invoice getInvoiceById(Long id) {
-
-        Invoice found = this.invoiceRepository.findById(id)
-                .orElseThrow(() -> new AppException(Error.INVOICE_NOT_FOUND));
-        return found;
-    }
 
     @Override
     @Transactional
-    public InvoiceWithItemsDto addNewInvoice(InvoicePostDto invoicePostDto) {
+    public InvoiceWithItemsDto addNewInvoice(InvoicePostDto invoicePostDto, AppUser authUser) {
 
-        Invoice mappedInvoice = new Invoice(invoicePostDto.getDescription(), invoicePostDto.getPaymentDue(), invoicePostDto.getItems());
+        Invoice mappedInvoice = new Invoice(invoicePostDto.getDescription(), invoicePostDto.getPaymentDue(),
+                invoicePostDto.getItems(), authUser);
         for (Item item : mappedInvoice.getItems()) {
             item.setInvoice(mappedInvoice);
         }
@@ -68,7 +64,7 @@ public class InvoiceServiceImpl implements InvoiceService {
 
     @Override
     @Transactional
-    public InvoiceWithItemsDto updateInvoice(InvoicePutDto updatedInvoice) {
+    public InvoiceWithItemsDto updateInvoice(InvoicePutDto updatedInvoice, AppUser appUser) {
         Optional<Invoice> exists = this.invoiceRepository.findById(updatedInvoice.getId());
 
         if (exists.isEmpty()) {
@@ -76,6 +72,11 @@ public class InvoiceServiceImpl implements InvoiceService {
         }
 
         Invoice invoice = exists.get();
+
+        if(invoice.getAppUser().getId() != appUser.getId()){
+            throw new AppException(Error.FORBIDDEN);
+        }
+
         invoice.setDescription(updatedInvoice.getDescription());
         invoice.setPaymentDue(updatedInvoice.getPaymentDue());
         invoice.setItems(updatedInvoice.getItems());
@@ -90,17 +91,24 @@ public class InvoiceServiceImpl implements InvoiceService {
 
     @Override
     @Transactional
-    public void deleteInvoice(Long invoiceId) {
+    public void deleteInvoice(Long invoiceId, AppUser authUser) {
         Invoice invoice = this.invoiceRepository.findById(invoiceId).orElseThrow(() ->
                 new AppException(Error.INVOICE_NOT_FOUND));
+
+        if(invoice.getAppUser().getId() != authUser.getId()){
+            throw new AppException(Error.FORBIDDEN);
+        }
         this.invoiceRepository.delete(invoice);
     }
 
     @Override
-    public InvoiceWithItemsDto getInvoiceWithItems(Long id) {
+    public InvoiceWithItemsDto getInvoiceWithItems(Long id, AppUser authUser) {
         Invoice invoice = this.invoiceRepository.findById(id)
                 .orElseThrow(() -> new AppException(Error.INVOICE_NOT_FOUND));
 
+        if(invoice.getAppUser().getId() != authUser.getId()){
+            throw new AppException(Error.FORBIDDEN);
+        }
         return mapInvoiceToInvoiceWithItemsDto(invoice);
     }
